Split up `ComponentInterface` module into multiple files.
Fixes #286.

The `ComponentInterface` struct and all its various member structs
have been getting far too large to live in a single source file.
This commit attempts to split them up into logical manageable chunks
with only a few structs per file.

This is a *big* patch, but it's almost entirely mechanical, and I've
avoided the temptation to perform any code changes beyond just moving
things into separate files. There are probably some smaller opportunities
for cleanup that have been surfaced here, but they can be for followups.

I've also taken the opportunity to add a lot more docs and a lot more
tests, which I think are much more tractable now that the code is split
up into smaller chunks.
